Implement authorization and roles permission management

+++ app/controllers/profiles/accounts_controller.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
 class Profiles::AccountsController < ApplicationController
+  before_action :authenticate_user!
+  before_action :set_user, except: %i[unlink]
+
   def show; end
 
   def unlink
@@ -9,8 +12,17 @@ class Profiles::AccountsController < ApplicationController
 
     return render_404 unless identity
 
+    authorize :identity_provider, :unlink?
+
     identity.destroy
 
     redirect_to profile_account_path
+  end
+
+  private
+
+  def set_user
+    @user = current_user
+    authorize @user
   end
 end
